Update KTLOSAQ40Loot.json
Updates from R2 AQ40. Also added Beeny to the list for next T2.5 chest, he hasnt gotten anything. and removed pennington from prio on all the caster loot as he was in R1 and he would have prio on every single caster item in R2 if it were not changed since nobody in R2 is on LC for caster items. He got cloak tonight from that..